refactor(torii): remove component storage relationship
