+++ crates/torii/src/tests/mod.rs
@@ -1,4 +1,4 @@
 mod common;
-mod components_test;
 mod entities_test;
 mod events_test;
+mod storages_test;
